Style: FIrebase관련 Code중첩 처리
